many changes to make clang work on windows
